✨ feat #71: 관리자 계정 열람, 수정, 삭제 기능 추가

+++ src/main/java/com/example/devlife/controller/BoardController.java
@@ -2,10 +2,7 @@ package com.example.devlife.controller;
 
 import com.example.devlife.dto.BoardResponseDto;
 import com.example.devlife.dto.BoardWriteRequestDto;
-import com.example.devlife.entity.Category;
-import com.example.devlife.entity.Comment;
-import com.example.devlife.entity.Grade;
-import com.example.devlife.entity.User;
+import com.example.devlife.entity.*;
 import com.example.devlife.repository.user.UserRepository;
 import com.example.devlife.service.board.BoardService;
 import com.example.devlife.service.comment.CommentService;
@@ -68,7 +65,7 @@ public class BoardController {
 			model.addAttribute("account", user);
 		}
 
-        if(user.getGrade().ordinal() < result.getGrade().ordinal() && (!Objects.equals(user.getProviderId(), result.getUser().getProviderId())) ) {
+        if(user.getGrade().ordinal() < result.getGrade().ordinal() && (!Objects.equals(user.getProviderId(), result.getUser().getProviderId()))  && (user.getRole() == Role.USER)  ) {
             return "board/YouShallNotPass";
         }
         else{
@@ -81,7 +78,7 @@ public class BoardController {
     public String boardUpdateForm(@PathVariable Long id, Model model,
                                   @AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : account") User account) {
         BoardResponseDto result = boardService.boardDetail(id);
-        if (!Objects.equals(result.getUser().getProviderId(), account.getProviderId())) {
+        if (!Objects.equals(result.getUser().getProviderId(), account.getProviderId()) && (account.getRole() == Role.USER) ) {
             return "board/YouShallNotPass";
         }
 
@@ -114,7 +111,7 @@ public class BoardController {
 
         BoardResponseDto result = boardService.boardDetail(id);
 
-        if (Objects.equals(result.getUser().getProviderId(), account.getProviderId())) {
+        if (Objects.equals(result.getUser().getProviderId(), account.getProviderId()) && (account.getRole() == Role.USER)) {
             boardService.deleteBoard(id);
             return "redirect:/board/list";
         }else{
